requirements.txt
requests
yt-dlp
python-dotenv
streamlit
langchain
langchain-google-genai
